requirements.txt
openai>=0.11.0
pandas>=1.1.0
pydantic>=1.7.0
Faker>=8.0.0
python-docx>=0.8.10
reportlab>=3.5.0